# Eliminate array copies