--- a/bmq-sdk/src/main/java/com/bloomberg/bmq/impl/QueueImpl.java
+++ b/bmq-sdk/src/main/java/com/bloomberg/bmq/impl/QueueImpl.java
@@ -265,9 +265,7 @@ public class QueueImpl implements QueueHandle {
     }
 
     public void pack(PutMessageImpl message) throws BMQException {
-        synchronized (lock) {
-            putMessages.get().add(message);
-        }
+        putMessages.get().add(message);
     }
 
     public void flush() throws BMQException {