--- a/bmq-sdk/src/test/java/com/bloomberg/bmq/impl/infr/proto/AckEventImplBuilderTest.java
+++ b/bmq-sdk/src/test/java/com/bloomberg/bmq/impl/infr/proto/AckEventImplBuilderTest.java
@@ -74,8 +74,7 @@ public class AckEventImplBuilderTest {
             builder.packMessage(m);
         }
 
-        ByteBuffer[] message;
-        message = builder.build();
+        ByteBuffer[] message = builder.build();
 
         TestHelpers.compareWithFileContent(message, MessagesTestSamples.ACK_MSG);
     }