*Issue number of the reported bug or feature request: #30 

**Describe your changes**
1. BrokerSession passes Collection<PutMessageImpl> instead of PutMessageImpl[] to avoid making copies.
2. ApplicationData calculates crc32c on the fly to avoid having to make an expensive copy of the whole payload.
3. ByteBufferOutputStream:
   - add peek() call to make an independent view of the underlying data.
   - don't copy and repack ByteBuffer's written to the ByteBufferOutputStream.
4. NettyTcpConnection still makes a copy of data read, but let netty make the copy instead of us.

**Testing performed**
Unit and integration tests have been updated to reflect the changes.

**Additional context**
As discussed with @sgalichkin